Breaking out the base class (and make one, where non exists) into a separate header file breaks the circular dependency.
