+++ src/Node.ipp
@@ -6,6 +6,15 @@
  */
 
 /**************************************************************************************
+ * INCLUDE
+ **************************************************************************************/
+
+#include "Publisher.hpp"
+#include "Subscription.hpp"
+#include "ServiceClient.hpp"
+#include "ServiceServer.hpp"
+
+/**************************************************************************************
  * PUBLIC MEMBER FUNCTIONS
  **************************************************************************************/
