+++ src/ServiceClientBase.hpp
@@ -5,8 +5,8 @@
  * Contributors: https://github.com/107-systems/107-Arduino-Cyphal/graphs/contributors.
  */
 
-#ifndef INC_107_ARDUINO_CYPHAL_CANARDSUBSCRIPTION_H
-#define INC_107_ARDUINO_CYPHAL_CANARDSUBSCRIPTION_H
+#ifndef INC_107_ARDUINO_CYPHAL_SERVICE_CLIENT_BASE_HPP
+#define INC_107_ARDUINO_CYPHAL_SERVICE_CLIENT_BASE_HPP
 
 /**************************************************************************************
  * INCLUDE
@@ -14,7 +14,7 @@
 
 #include <memory>
 
-#include "libcanard/canard.h"
+#include "SubscriptionBase.h"
 
 /**************************************************************************************
  * NAMESPACE
@@ -27,31 +27,13 @@ namespace impl
  * CLASS DECLARATION
  **************************************************************************************/
 
-class SubscriptionBase
+template <typename T_REQ>
+class ServiceClientBase : public SubscriptionBase
 {
 public:
-  SubscriptionBase(CanardTransferKind const transfer_kind)
-  : _transfer_kind{transfer_kind}
-  {
-    _canard_rx_sub.user_reference = static_cast<void *>(this);
-  }
-  virtual ~SubscriptionBase() { }
-  SubscriptionBase(SubscriptionBase const &) = delete;
-  SubscriptionBase(SubscriptionBase &&) = delete;
-  SubscriptionBase &operator=(SubscriptionBase const &) = delete;
-  SubscriptionBase &operator=(SubscriptionBase &&) = delete;
-
-
-  virtual bool onTransferReceived(CanardRxTransfer const & transfer) = 0;
-
-
-  [[nodiscard]] CanardTransferKind canard_transfer_kind() const { return _transfer_kind; }
-  [[nodiscard]] CanardRxSubscription &canard_rx_subscription() { return _canard_rx_sub; }
-
-
-private:
-  CanardTransferKind const _transfer_kind;
-  CanardRxSubscription _canard_rx_sub;
+  ServiceClientBase() : SubscriptionBase{CanardTransferKindResponse} { }
+  virtual ~ServiceClientBase() { }
+  virtual bool request(CanardNodeID const remote_node_id, T_REQ const & req) = 0;
 };
 
 /**************************************************************************************
@@ -64,7 +46,7 @@ private:
  * TYPEDEF
  **************************************************************************************/
 
-using Subscription = std::shared_ptr<impl::SubscriptionBase>;
+template <typename T_REQ>
+using ServiceClient = std::shared_ptr<impl::ServiceClientBase<T_REQ>>;
 
-
-#endif /* INC_107_ARDUINO_CYPHAL_CANARDSUBSCRIPTION_H */
+#endif /* INC_107_ARDUINO_CYPHAL_SERVICE_CLIENT_BASE_HPP */
