+++ src/ServiceServerBase.hpp
@@ -5,8 +5,8 @@
  * Contributors: https://github.com/107-systems/107-Arduino-Cyphal/graphs/contributors.
  */
 
-#ifndef INC_107_ARDUINO_CYPHAL_CANARDSUBSCRIPTION_H
-#define INC_107_ARDUINO_CYPHAL_CANARDSUBSCRIPTION_H
+#ifndef INC_107_ARDUINO_CYPHAL_SERVICE_SERVER_BASE_HPP
+#define INC_107_ARDUINO_CYPHAL_SERVICE_SERVER_BASE_HPP
 
 /**************************************************************************************
  * INCLUDE
@@ -14,7 +14,7 @@
 
 #include <memory>
 
-#include "libcanard/canard.h"
+#include "SubscriptionBase.h"
 
 /**************************************************************************************
  * NAMESPACE
@@ -27,31 +27,10 @@ namespace impl
  * CLASS DECLARATION
  **************************************************************************************/
 
-class SubscriptionBase
+class ServiceServerBase : public SubscriptionBase
 {
 public:
-  SubscriptionBase(CanardTransferKind const transfer_kind)
-  : _transfer_kind{transfer_kind}
-  {
-    _canard_rx_sub.user_reference = static_cast<void *>(this);
-  }
-  virtual ~SubscriptionBase() { }
-  SubscriptionBase(SubscriptionBase const &) = delete;
-  SubscriptionBase(SubscriptionBase &&) = delete;
-  SubscriptionBase &operator=(SubscriptionBase const &) = delete;
-  SubscriptionBase &operator=(SubscriptionBase &&) = delete;
-
-
-  virtual bool onTransferReceived(CanardRxTransfer const & transfer) = 0;
-
-
-  [[nodiscard]] CanardTransferKind canard_transfer_kind() const { return _transfer_kind; }
-  [[nodiscard]] CanardRxSubscription &canard_rx_subscription() { return _canard_rx_sub; }
-
-
-private:
-  CanardTransferKind const _transfer_kind;
-  CanardRxSubscription _canard_rx_sub;
+  ServiceServerBase() : SubscriptionBase{CanardTransferKindRequest} { }
 };
 
 /**************************************************************************************
@@ -64,7 +43,6 @@ private:
  * TYPEDEF
  **************************************************************************************/
 
-using Subscription = std::shared_ptr<impl::SubscriptionBase>;
+using ServiceServer = std::shared_ptr<impl::ServiceServerBase>;
 
-
-#endif /* INC_107_ARDUINO_CYPHAL_CANARDSUBSCRIPTION_H */
+#endif /* INC_107_ARDUINO_CYPHAL_SERVICE_SERVER_BASE_HPP */
