+++ src/Subscription.hpp
@@ -12,17 +12,9 @@
  * INCLUDE
  **************************************************************************************/
 
-#include <memory>
-
 #include "SubscriptionBase.h"
 
-#include "libcanard/canard.h"
-
-/**************************************************************************************
- * FORWARD DECLARATION
- **************************************************************************************/
-
-class Node;
+#include "Node.hpp"
 
 /**************************************************************************************
  * NAMESPACE
@@ -62,12 +54,6 @@ private:
  **************************************************************************************/
 
 } /* impl */
-
-/**************************************************************************************
- * TYPEDEF
- **************************************************************************************/
-
-using Subscription = std::shared_ptr<impl::SubscriptionBase>;
 
 /**************************************************************************************
  * TEMPLATE IMPLEMENTATION
